MAINT: corrcoef, memory usage optimization
We calculate sqrt on the small vector rather than on that huge
product matrix and we combine the "outer" product with element-wise
devision. So even though we have a slower loop over the rows now ... this
code snipped runs about 3 times faster than before.

However the speed improvement of the whole function is not really
significant because cov() takes 80-99% of the time (dependent on
blas/lapack implementation and number of CPU cores).

More important is that we will safe 1/3 memory. For example corrcoef()
for a [23k, m]  matrix needs 8GB now instead of 12GB.